requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==4.4.2
Deprecated==1.2.12
idna==2.10
networkx==2.5.1
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
python-dotenv==0.18.0
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.4
wrapt==1.12.1
